+++ main.go
@@ -2,35 +2,20 @@ package main
 
 import (
 	"fmt"
-	"github.com/Mowinski/stx-next-go-workshop/vies"
-	"io"
 	"os"
+
+	"github.com/Mowinski/stx-next-go-workshop/vatno"
 )
 
-func readFrom(reader io.Reader) (data vies.InputData, err error) {
-	fmt.Print("Enter country code: ")
-	_, err = fmt.Fscanf(reader, "%s", &data.CountryCode)
-	if err != nil {
-		return data, err
-	}
-
-	fmt.Print("Enter VAT: ")
-	_, err = fmt.Fscanf(reader, "%s", &data.VATNumber)
-	if err != nil {
-		return data, err
-	}
-
-	return data, nil
-}
-
-func writeTo(writer io.Writer, vatDetail *vies.VatDetails) {
-
+func abort(err error) {
+	fmt.Println(err)
+	os.Exit(1)
 }
 
 func main() {
-	inputData, err := readFrom(os.Stdin)
+	vatNo, err := vatno.Read(os.Stdin)
 	if err != nil {
-		panic(err)
+		abort(err)
 	}
-	fmt.Println(inputData)
+	fmt.Println(vatNo)
 }
